[Evals] 1/N Add more task related tests (#110)
# What does this PR do?

Adds more task specific tests. The tests are pretty lightweight and basic. The idea is to be able to catch any errors in templating and having basic E2E correctness checks for each task to avoid issues like #108 

There is some repetition across tests which is intentional: It's best to decouple tests for different tasks. 

I'll add more tests for other tasks in future PRs. 
---------

Signed-off-by: SumanthRH <sumanthrh@anyscale.com>